Dialogs and Popups
Adds a demo to open popups and dialogs programatically.

Popups are visible, we have to figure out why it doesn’t work within
UXP.

Dialogs do work if the `.XamlRoot(...)` property is set manually. Having
a `ContentDialog` as part of a class leads to a null pointer dereference
when destructing it though.

See relevant code comments in `DialogsAndPopupsDemo.cpp`.